ai-service/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0
# Provider SDKs (optional for stubs; wire later as needed)
groq>=0.9.0
edge-tts>=6.1.12
# whisper is heavy; install only when needed
# openai-whisper>=20231117
